Use the application mainpage for alert dialogs
Removed the IPageAware interface from the PageDialogService as it
wasn’t needed. All methods now make use of the Application’s MainPage
to show Alerts and ActionSheets. There is no longer a need for the Page
property in the service.

Also reverted the previous commit because of the fragility it adds.

fixes #254