Add #define DEBUG to DebugLogger
The DebugLogging class does not do anything useful unless compiled using DEBUG mode. Nuget packages are compiled in release mode which only leaves the string formating code without doing anything with the result and if a developer try to use this class from their application code it will be confusing and misleading since it actually does not log anything.