included tests for each function changed + considered enum name or value during conversion
Allow bind Enum values in IAutoInitialize ViewModels

Currently if anyone try pass through a enum value as raw type(value or name) within INavigationParameters get an error of conversion, to solve this problem was included the conditions above.

Allow bind Enum values in IAutoInitialize ViewModels

Currently if anyone try pass through a enum value as raw type(value or name) within INavigationParameters get an error of conversion, to solve this problem was included the conditions above.

Update ParametersExtensions.cs

force to use only enum name

I'm considering to use only the enum name because if the value is being considered during the conversion can cause side effect because the value as number can be refered to a lot of things in navigation parameters not the enum type searched.