Reword description of INavigationResult
The wording before suggested that if the navigation request was
unsuccessful there would be an exception captured; however, this is not
the case when navigation is attempted and CanNavigate is false. In this
case, Success would be false and Exception would be null.

The wording now does not make that suggestion.